refactor: new implementation for sending key combinations
The `IsNeedKeep()` (with tick fault tolerance) introduced in the previous version will expose the issue in #69. Because when the user sends a right click, the program captures the middle click operation sent by itself, causing the two operations to be within the tick fault tolerance range, thus triggering the sending of the two commands in `HandleMiddleClick`. To this end, this commit introduces `dwExtraInfo` information to prevent the capture of operations sent by the program itself.