build(workflows): support automatic build and on-demand release (#116)
This commit supports automatically building an alpha version with the latest short hash and uploading it to Artifacts when there is a push on the main branch.

It also supports manually running build.yml, which will require entering a version number that complies with semantic versioning. It will automatically build and create a PR that includes a commit updating the version number in version.h and submits the corresponding binary files to setdll. When this PR is merged, since version.h is updated, it will automatically trigger the generation of release notes based on the latest version number and complete a series of release operations.

---------

Co-authored-by: YorkWaugh <67750590+YorkWaugh@users.noreply.github.com>