fix(iaccessible): nullptr reference in `IsNameNewTab`
Users report that when upgrading to Chrome 143.0.7499.41, it crashes when clicking on web pages.

The dmp file shows that the crash happens in `MouseProc` when dealing with `wcslen`. The only possible cause is that the function does not find the new tab button with an expected name but we try to calculate its length anyway. It is weird since it should always find it, and it does work on our test machines. Maybe Chrome changes something in certain configurations and we may need to figure out later. Add a nullptr check to avoid the crash and let it fallback to other methods for now.

Besides, this function should not be called when clicking on web pages (which is not bookmarks at all), we will improve that in a later commit.

Hopefully fix #191.