fix(iaccessible): fix double click close failure by pruning a11y traversal
This fixes a regression in which we replace `FindPageTabList` with `FindElementWithRole`. However, `FindElementWithRole` uses an unbounded deepth-first search which may traverse into web content instead of staying within the browser UI tree. `PtInRect` may fail because it finds an incorrect node with coincidentally the expected role.

We reverts it back to `FindPageTabList`, which restricts the search to children with either `ROLE_SYSTEM_PANE` or `ROLE_SYSTEM_TOOLBAR`.

Fix #199.